clean up legacy csl parser and move it to utils since it's still being used
