+++ src/js/_legacy/_entrypoints/options-page/__tests__/style-form-test.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import toJSON from 'enzyme-to-json';
 import { shallow } from 'enzyme';
 
-import PCSL from '_legacy/utils/parse-csl';
+import { parseCSL as PCSL } from 'utils/file';
 import StyleForm from '../style-form';
 
 const parseCSL = PCSL as jest.Mock;
